feat;added home link to nav
